Refactor store profile with banking info, enhance auth flow, and add product deletion
- Introduce `BankingInfo` DTO and restructure `StoreDTO` to encapsulate bank details, updating `StoreViewModel`, `StoreState`, and `StoreInfoScreen` accordingly
- Enhance authentication flow: extend `LoginSuccess` event with `hasStore` flag, differentiate between unregistered and incomplete store profiles in `LoginScreen` with tailored dialog messages
- Implement product deletion: add `deleteProduct` to `ProductDetailViewModel`, confirm dialog in `ProductDetailScreen`, and long-press delete action in `ProductGridItem` with `ProductUiEvent` for refresh/toast feedback
- Refactor category deletion: change API to `@PUT`, add local `deleteCategoryById`, return `Result<Unit>` from repository, and show result dialogs in `CategoryViewModel`
- Standardize price handling: change `formatPrice` to accept `Double`, add `formatPriceToDouble` parser, and update all call sites (checkout, scanner, order, product) to use `toDouble()`
- Externalize API base URL: move to `gradle.properties`, generate `BuildConfig.BASE_URL`, and consume in `NetworkModule` and `CheckoutViewModel`
- Update `UserDTO` with separate `storeFromGetMe`/`storeFromLogin` fields and computed `store` property

+++ app/src/main/java/com/duongnd/pocketposapp/feature/checkout/components/ReceiptCard.kt
@@ -142,7 +142,7 @@ fun ReceiptCard(
                     fontWeight = FontWeight.ExtraBold
                 )
                 Text(
-                    formatPrice(totalPrice) + " đ",
+                    formatPrice(totalPrice.toDouble()) + " đ",
                     style = MaterialTheme.typography.headlineSmall,
                     fontWeight = FontWeight.ExtraBold,
                     color = primaryColor
